+++ R/netgraph.netmeta.R
@@ -288,28 +288,33 @@
 #' # Network graph optimized, starting from a circle, with multi-arm
 #' # study colored
 #' #
-#' netgraph(net1, start = "circle", iterate = TRUE, col.multiarm = "purple")
+#' netgraph(net1, start = "circle", iterate = TRUE,
+#'          multiarm = TRUE, col.multiarm = "purple")
 #'
 #' # Network graph optimized, starting from a circle, with multi-arm
 #' # study colored and all intermediate iteration steps visible
 #' #
-#' netgraph(net1, start = "circle", iterate = TRUE, col.multiarm = "purple",
+#' netgraph(net1, start = "circle", iterate = TRUE,
+#'          multiarm = TRUE, col.multiarm = "purple",
 #'          allfigures = TRUE)
 #' 
 #' # Network graph optimized, starting from Laplacian eigenvectors,
 #' # with multi-arm study colored
 #' #
-#' netgraph(net1, start = "eigen", col.multiarm = "purple")
+#' netgraph(net1, start = "eigen",
+#'          multiarm = TRUE, col.multiarm = "purple")
 #' 
 #' # Network graph optimized, starting from different Laplacian
 #' # eigenvectors, with multi-arm study colored
 #' #
-#' netgraph(net1, start = "prcomp", col.multiarm = "purple")
+#' netgraph(net1, start = "prcomp",
+#'          multiarm = TRUE, col.multiarm = "purple")
 #' 
 #' # Network graph optimized, starting from random initial layout,
 #' # with multi-arm study colored
 #' #
-#' netgraph(net1, start = "random", col.multiarm = "purple")
+#' netgraph(net1, start = "random",
+#'          multiarm = TRUE, col.multiarm = "purple")
 #' 
 #' # Network graph without plastic look and one highlighted comparison
 #' #
@@ -354,7 +359,7 @@ netgraph.netmeta <- function(x, seq = x$seq,
                              highlight = NULL, col.highlight = "red2",
                              scale.highlight = 1,
                              ##
-                             multiarm = any(x$narms > 2),
+                             multiarm = FALSE,
                              col.multiarm = NULL,
                              alpha.transparency = 0.5,
                              ##
@@ -599,8 +604,14 @@ netgraph.netmeta <- function(x, seq = x$seq,
 
   if (missing(thickness)) {
     if (start.layout == "circle" & iterate == FALSE & plastic == TRUE) {
-      thick <- "se.fixed"
-      thickness <- "se.fixed"
+      if (x$comb.random & !x$comb.fixed) {
+        thick <- "se.random"
+        thickness <- "se.random"
+      }
+      else {
+        thick <- "se.fixed"
+        thickness <- "se.fixed"
+      }
     }
     else {
       thick <- "equal"
@@ -883,10 +894,9 @@ netgraph.netmeta <- function(x, seq = x$seq,
            "number of treatments.",
            eval. = FALSE)
     if (is.null(names(srt.labels)))
-      dat.nodes$srt <- srt.labels[seq1]
+      dat.nodes$srt <- srt.labels
     else {
       ## Check names of named vector 'srt.labels'
-      srt.labels <- srt.labels[seq1]
       names.srt.labels <- names(srt.labels)
       names.srt.labels <- setseq(names.srt.labels, dat.nodes$trts,
                                  paste0("Names of vector provided in ",